+++ packages/dialect-react-sdk/hooks/useDapp.ts
@@ -1,0 +1,35 @@
+import type { Dapp, DialectSdkError } from '@dialectlabs/sdk';
+import useSWR from 'swr';
+import { EMPTY_OBJ } from '../utils';
+import { DAPP_CACHE_KEY } from './internal/swrCache';
+import useDialectSdk from './useDialectSdk';
+
+interface UseDappValue {
+  dapp: Dapp | null;
+  isFetching: boolean;
+  errorFetching: DialectSdkError | null;
+}
+
+interface UseDappParams {
+  refreshInterval?: number;
+}
+
+function useDapp({ refreshInterval }: UseDappParams = EMPTY_OBJ): UseDappValue {
+  const { dapps } = useDialectSdk();
+  const {
+    info: { wallet },
+  } = useDialectSdk();
+  const { data: dapp, error } = useSWR(
+    DAPP_CACHE_KEY(wallet),
+    () => dapps.find(),
+    { refreshInterval, refreshWhenOffline: true }
+  );
+
+  return {
+    dapp: dapp || null,
+    isFetching: !error && dapp === undefined,
+    errorFetching: error,
+  };
+}
+
+export default useDapp;