+++ packages/dialect-react-ui/entities/wallet-states/EncryptionInfo.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
-import { Centered, Loader } from '../../common';
-import { H3, P } from '../../common/preflighted';
-import { useTheme } from '../../common/providers/DialectThemeProvider';
+import { Centered, Loader } from '../../components/common';
+import { H3, P } from '../../components/common/preflighted';
+import { useTheme } from '../../components/common/providers/DialectThemeProvider';
 
 const EncryptionInfo = () => {
   const { textStyles } = useTheme();
@@ -12,7 +12,7 @@ const EncryptionInfo = () => {
       >
         <Loader className="dt-mr-2" /> Waiting for your wallet
       </H3>
-      <P className="dt-text-center dt-max-w-sm dt-min-w-[24rem] dt-opacity-50">
+      <P className="dt-text-center dt-max-w-sm dt-opacity-50">
         Please create encryption keys to be able to read and send encrypted
         messages. It is free and does not involve the network.
       </P>