+++ packages/dialect-react-sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [UNRELEASED]
 
+- feature: add `useDapp` and `useDappAddresses` hooks to the `react-sdk`
+
 ## [1.0.0-beta.5] - 2022-06-30
 
 - chore: bump @dialectlabs/sdk